+++ axelor-human-resource/src/main/java/com/axelor/apps/hr/service/timesheet/timer/TimesheetTimerService.java
@@ -1,12 +1,13 @@
 package com.axelor.apps.hr.service.timesheet.timer;
 
 import com.axelor.apps.hr.db.TSTimer;
+import com.axelor.apps.hr.db.TimesheetLine;
 
 public interface TimesheetTimerService {
 
 	public void pause(TSTimer timer);
 	public void stop(TSTimer timer);
 	public void calculateDuration(TSTimer timer);
-	public void generateTimesheetLine(TSTimer timer);
+	public TimesheetLine generateTimesheetLine(TSTimer timer);
 	public TSTimer getCurrentTSTimer();
 }
